baseboard/brya/relevant-paths.txt firmware-brya-14505.B-main

Relevant changes:

git log --oneline 043c2a0a54..aca2261e16 -- baseboard/brya board/agah
board/anahera board/banshee board/brya board/crota board/felwinter
board/gimble board/kano board/mithrax board/osiris board/primus
board/redrix board/taeko board/taniks board/vell board/volmar
driver/bc12/pi3usb9201_public.* driver/charger/bq25710.*
driver/ppc/nx20p348x.* driver/ppc/syv682x_public.*
driver/retimer/bb_retimer_public.* driver/tcpm/nct38xx.*
driver/tcpm/ps8xxx_public.* driver/tcpm/tcpci.* include/power/alderlake*
include/intel_x86.h power/alderlake* power/intel_x86.c
util/getversion.sh

aca2261e16 volmar: Prevent discharging when battery is disconnected
e9314be72f Zephyr: Add shim support for PS8818
ae28b6ef3a volmar: Add new manufacturer name for battery AP20CBL
6278fe9df7 USB Mux: Unify drivers for safe mode
77535ac0c2 osiris: Discharge on AC when high MLCC noise
4a85adc328 Move Brya PPCs into the devicetree
ded9307b79 util/getversion.sh: Fix version when not in a git repo
a29d4c0737 brya: Config-guard TBT UFP code
f129262606 mithrax: Remove Burnside Bridge config
23205994cc brya: usb_mux: Split struct usb_mux in brya boards
71b2ef709d Update license boilerplate text in source code files
538a24ab36 config: rename HOSTCMD_ESPI_* to HOST_INTERFACE_ESPI_*

BRANCH=None
BUG=b:246214501 b:245042158 b:218976941 b:236131108 b:236274003
BUG=b:244467636 b:195416058 chromium:1098010 b:246424843 b:244457125
BUG=b:234772074
TEST=`make -j buildall`

Force-Relevant-Builds: all

+++ common/fpsensor/fpsensor.c
@@ -1,4 +1,4 @@
-/* Copyright 2017 The Chromium OS Authors. All rights reserved.
+/* Copyright 2017 The ChromiumOS Authors
  * Use of this source code is governed by a BSD-style license that can be
  * found in the LICENSE file.
  */
@@ -481,7 +481,11 @@ static enum ec_status fp_command_frame(struct host_cmd_handler_args *args)
 				FP_CONTEXT_ENCRYPTION_SALT_BYTES);
 		trng_exit();
 
-		if (fgr == template_newly_enrolled) {
+		/*
+		 * TODO(http://b/244781166): Use consistent types so cast is
+		 * not needed.
+		 */
+		if (fgr == (uint32_t)template_newly_enrolled) {
 			/*
 			 * Newly enrolled templates need new positive match
 			 * salt, new positive match secret and new validation
